+++ src/main/kotlin/org/nemwiz/jiracommitmessage/configuration/PrefixType.kt
@@ -1,0 +1,8 @@
+package org.nemwiz.jiracommitmessage.configuration
+
+enum class PrefixType(val type: String) {
+    NO_PREFIX("No prefix"),
+    DASH("-"),
+    COLON(":"),
+    HASH("#"),
+}