+++ src/main/kotlin/org/nemwiz/jiracommitmessage/services/JiraCommitMessagePlugin.kt
@@ -3,14 +3,13 @@ package org.nemwiz.jiracommitmessage.services
 import com.intellij.notification.BrowseNotificationAction
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
-import org.nemwiz.jiracommitmessage.configuration.InfixType
-import org.nemwiz.jiracommitmessage.configuration.MessageWrapperType
 import org.nemwiz.jiracommitmessage.configuration.PluginSettingsState
 import org.nemwiz.jiracommitmessage.provider.PluginNotifier
 import java.util.*
 import java.util.regex.Pattern
 
 private const val DEFAULT_REGEX_FOR_JIRA_PROJECT_ISSUES = "([A-Z]+[_-][0-9]+)"
+private const val CONVENTIONAL_COMMITS_REGEX = "(feat|fix|build|ci|chore|docs|perf|refactor|style|test)"
 
 class JiraCommitMessagePlugin(private val project: Project) : Disposable {
 
@@ -38,14 +37,21 @@ class JiraCommitMessagePlugin(private val project: Project) : Disposable {
             return ""
         }
 
+        val isConventionalCommit = PluginSettingsState.instance.state.isConventionalCommit
+
         val jiraIssue = extractJiraIssueFromBranch(isAutoDetectProjectKey, branchName, jiraProjectKeys)
+        val conventionalCommitType = extractConventionalCommitType(isConventionalCommit, branchName)
 
         val selectedMessageWrapper = PluginSettingsState.instance.state.messageWrapperType
+        val selectedPrefixType = PluginSettingsState.instance.state.messagePrefixType
         val selectedInfixType = PluginSettingsState.instance.state.messageInfixType
 
-        return if (jiraIssue != null)
-            createCommitMessage(selectedMessageWrapper, selectedInfixType, jiraIssue)
-        else ""
+        return CommitMessageBuilder(jiraIssue)
+            .withWrapper(selectedMessageWrapper)
+            .withInfix(selectedInfixType)
+            .withConventionalCommit(conventionalCommitType)
+            .withPrefix(selectedPrefixType)
+            .getCommitMessage()
     }
 
     private fun extractJiraIssueFromBranch(
@@ -75,35 +81,14 @@ class JiraCommitMessagePlugin(private val project: Project) : Disposable {
         return jiraIssue
     }
 
-    private fun createCommitMessage(wrapperType: String, infixType: String, jiraIssue: String): String {
-        val messageWithWrapper = addWrapper(wrapperType, jiraIssue)
-        return addInfix(infixType, messageWithWrapper)
+    private fun extractConventionalCommitType(isConventionalCommit: Boolean, branchName: String): String? {
+        if (isConventionalCommit) {
+            val pattern = Pattern.compile(CONVENTIONAL_COMMITS_REGEX).toRegex()
+            val matchedConventionalType = pattern.find(branchName)
+            return matchedConventionalType?.value
+        }
+        return null
     }
-
-    private fun addWrapper(wrapperType: String, jiraIssue: String) =
-        if (wrapperType == MessageWrapperType.NO_WRAPPER.type) {
-            jiraIssue
-        } else {
-            String.format(
-                Locale.US,
-                "%s%s%s",
-                wrapperType.substring(0, 1),
-                jiraIssue,
-                wrapperType.substring(1, 2)
-            )
-        }
-
-    private fun addInfix(infixType: String, commitMessage: String) =
-        if (infixType == InfixType.NO_INFIX.type) {
-            commitMessage
-        } else {
-            String.format(
-                Locale.US,
-                "%s%s",
-                commitMessage,
-                infixType
-            )
-        }
 
     override fun dispose() {
     }