+++ src/test/kotlin/org/nemwiz/jiracommitmessage/services/JiraCommitMessagePluginTest.kt
@@ -8,6 +8,7 @@ import io.mockk.verify
 import org.nemwiz.jiracommitmessage.configuration.InfixType
 import org.nemwiz.jiracommitmessage.configuration.MessageWrapperType
 import org.nemwiz.jiracommitmessage.configuration.PluginSettingsState
+import org.nemwiz.jiracommitmessage.configuration.PrefixType
 import org.nemwiz.jiracommitmessage.provider.PluginNotifier
 
 class JiraCommitMessagePluginTest : BasePlatformTestCase() {
@@ -40,7 +41,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1, dummyProjectKey2)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
 
-        val mockBranch1 = "feat/${dummyProjectKey2}-8172"
+        val mockBranch1 = "hello/${dummyProjectKey2}-8172"
 
         val plugin = project.service<JiraCommitMessagePlugin>()
 
@@ -48,7 +49,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
 
         assertEquals("(${dummyProjectKey2}-8172)", commitMessage1)
 
-        val mockBranch2 = "feat/${dummyProjectKey1}_123-foo"
+        val mockBranch2 = "hello/${dummyProjectKey1}_123-foo"
 
         val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
 
@@ -61,6 +62,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.CURLY.type
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
 
         val mockBranch = "this-is-an-awesome-branch-${dummyProjectKey1}-22"
 
@@ -80,6 +82,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
             listOf(dummyProjectKey1, dummyProjectKey2, dummyProjectKey3)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.NO_WRAPPER.type
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
 
         val mockBranch = "fix/${dummyProjectKey2}-381"
 
@@ -95,6 +98,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         val dummyProjectKey1 = "INFIX"
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.COLON.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.CURLY.type
 
@@ -119,6 +123,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         val dummyProjectKey1 = "FROG"
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
 
@@ -135,6 +140,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
 
         val dummyProjectKey1 = "FROG"
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
 
@@ -149,6 +155,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         val dummyProjectKey1 = "PRODUCT"
         PluginSettingsState.instance.state.jiraProjectKeys = emptyList()
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
 
@@ -175,6 +182,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         PluginSettingsState.instance.state.jiraProjectKeys =
             listOf(dummyProjectKey1, dummyProjectKey2, dummyProjectKey3)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
 
@@ -190,6 +198,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
     fun testReturnsEmptyStringWhenJiraProjectKeyCantBeAutomaticallyDetectedFromBranchName() {
         PluginSettingsState.instance.state.jiraProjectKeys = emptyList()
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
         PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
         PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
 
@@ -200,5 +209,160 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
 
         assertEquals("", commitMessage)
+    }
+
+    fun testReturnConventionalCommitInfixWhenConvetionalCommitsOptionIsOn() {
+        val dummyProjectKey1 = "LEGOS"
+
+        PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val mockBranch1 = "feat/awesome-LEGOS-123"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage1 = plugin.getCommitMessageFromBranchName(mockBranch1)
+
+        assertEquals("feat(${dummyProjectKey1}-123)", commitMessage1)
+
+        val mockBranch2 = "chore-awesome-LEGOS-123"
+
+        val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
+
+        assertEquals("chore(${dummyProjectKey1}-123)", commitMessage2)
+    }
+
+    fun testReturnConventionalCommitInfixWhenConvetionalCommitsOptionIsOnAndAutomaticDetectIsOn() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.COLON.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val mockBranch1 = "fix/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage1 = plugin.getCommitMessageFromBranchName(mockBranch1)
+
+        assertEquals("fix(PLUGIN-831):", commitMessage1)
+
+        val mockBranch2 = "docs-awesome-PLUGIN-831"
+
+        val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
+
+        assertEquals("docs(PLUGIN-831):", commitMessage2)
+    }
+
+    fun testDoesNotAddConventionalCommitInfixWhenConventionalCommitsAreNotOn() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = false
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.BOX.type
+
+        val mockBranch = "test/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
+
+        assertEquals("[PLUGIN-831]", commitMessage)
+    }
+
+    fun testOnlyAddsInfixFromTheSupportedConventionalCommitInfix() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val branches = listOf(
+            "test/PROJ-123",
+            "ci/PROJ-123",
+            "chore/PROJ-123",
+            "build/PROJ-123",
+            "fix/PROJ-123",
+            "feat/PROJ-123",
+            "docs/PROJ-123",
+            "perf/PROJ-123",
+            "refactor/PROJ-123",
+            "style/PROJ-123",
+            "unknown/PROJ-123",
+            "dummy/PROJ-123",
+        )
+
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessages = branches.map { b ->  plugin.getCommitMessageFromBranchName(b)}
+
+        assertEquals(branches.size, commitMessages.size)
+
+        assertFalse(commitMessages.contains("dummy(PROJ-123)"))
+        assertFalse(commitMessages.contains("unknown(PROJ-123)"))
+
+        assertTrue(commitMessages.contains("test(PROJ-123)"))
+        assertTrue(commitMessages.contains("ci(PROJ-123)"))
+        assertTrue(commitMessages.contains("chore(PROJ-123)"))
+        assertTrue(commitMessages.contains("build(PROJ-123)"))
+        assertTrue(commitMessages.contains("fix(PROJ-123)"))
+        assertTrue(commitMessages.contains("feat(PROJ-123)"))
+        assertTrue(commitMessages.contains("docs(PROJ-123)"))
+        assertTrue(commitMessages.contains("perf(PROJ-123)"))
+        assertTrue(commitMessages.contains("refactor(PROJ-123)"))
+        assertTrue(commitMessages.contains("style(PROJ-123)"))
+    }
+
+    fun testDoesNotAddPrefixWhenPrefixIsNotSet() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = false
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.NO_PREFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.BOX.type
+
+        val mockBranch = "test/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
+
+        assertEquals("[PLUGIN-831]", commitMessage)
+    }
+
+    fun testAddsPrefixWhenMessagePrefixIsSet() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = false
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.COLON.type
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.HASH.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.NO_WRAPPER.type
+
+        val mockBranch = "test/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
+
+        assertEquals("#PLUGIN-831:", commitMessage)
+    }
+
+    fun testAlwaysAddsPrefixToTheBeginningOfTheMessageWhenPrefixIsSet() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.pluginState.messagePrefixType = PrefixType.COLON.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.CURLY.type
+
+        val mockBranch = "test/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
+
+        assertEquals(":test{PLUGIN-831}", commitMessage)
     }
 }