+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Added
+- Support for conventional commits based on branch name
+- Ability to specify message prefix
+
 ## [3.2.0] - 2023-01-27
 
 ### Added